Adjust resolution data and improve image compression quality
Changes have been made to the screen resolution data that are passed by cookie. Now only the screen width is used, not the larger of width or height. Additionally, the compression quality for images has been sufficiently improved. A method has been added to the ImageCache class to automatically adjust image compression quality based on the image width.